+++ pages/admin/index.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
 import { NextPage } from "next";
 import Header from "@components/admin/Header";
-import Table from "@components/admin/Table";
+import Table from "@components/admin/ApplicationDashboardTable";
 import {
   ref,
   get,
